+++ apollo-gradle-plugin-incubating/src/main/resources/META-INF/gradle-plugins/com.apollographql.android.properties
@@ -1,1 +1,1 @@
-implementation-class=com.apollographql.apollo.gradle.ApolloPluginAndroid
+implementation-class=com.apollographql.apollo.gradle.internal.ApolloPluginAndroid
