+++ apollo-gradle-plugin-incubating/src/main/resources/META-INF/gradle-plugins/com.apollographql.apollo.properties
@@ -1,1 +1,1 @@
-implementation-class=com.apollographql.apollo.gradle.ApolloPlugin
+implementation-class=com.apollographql.apollo.gradle.internal.ApolloPlugin
